Entropic correction due to free energy profile  not constant for long urea-urea distances
The resulting free energy profile is not constant for long urea-urea distances. This is due to the probability of finding a urea molecule at a distance r is proportional to the area of a sphere 4πr2 . This purely entropic contribution to the free energy is −kT ln 4πr2 and has to be subtracted from the profile. Since we are only interested in relative free energy differences the correction is simplified to 2kT ln r.